+++ TaskExecutor/.idea/dictionaries/MaYP1998.xml
@@ -1,0 +1,7 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="MaYP1998">
+    <words>
+      <w>peng</w>
+    </words>
+  </dictionary>
+</component>
