+++ TaskExecutor/pom.xml
@@ -2,28 +2,46 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.github.mayp1998.taskExecutor</groupId>
-    <artifactId>taskExecutor</artifactId>
-    <version>1.0-SNAPSHOT</version>
+    <artifactId>TaskExecutor</artifactId>
+    <version>1.0-RELEASE</version>
+    <packaging>jar</packaging>
+    <name>TaskExecutor</name>
+    <description>Java universal task executor, base on the java.util.concurrent.ScheduledExecutorService.</description>
+    <url>https://github.com/MaYP1998/TaskExecutor.git</url>
+
+    <licenses>
+        <license>
+            <name>The Apache Software License, Version 2.0</name>
+            <url>http://www.apache.org/licenses/LICENSE-2.0.txt</url>
+            <distribution>repo</distribution>
+        </license>
+    </licenses>
+
+    <scm>
+        <url>https://github.com/MaYP1998/TaskExecutor.git</url>
+        <connection>https://github.com/MaYP1998/TaskExecutor.git</connection>
+        <developerConnection>https://github.com/MaYP1998/TaskExecutor.git</developerConnection>
+    </scm>
+
+
     <build>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.1</version>
+                <version>3.6.1</version>
                 <configuration>
-                    <source>8</source>
-                    <target>8</target>
+                    <source>1.8</source>
+                    <target>1.8</target>
                     <encoding>utf-8</encoding>
                 </configuration>
             </plugin>
         </plugins>
     </build>
-
-
-    <!--1 -->
 
     <distributionManagement>
         <repository>
@@ -40,13 +58,9 @@
         <profile>
             <id>disable-javadoc-doclint</id>
             <activation>
-                <jdk>[1.8,)</jdk>
+<!--                <jdk>[1.8,)</jdk>-->
+                <jdk>1.8</jdk>
             </activation>
-            <!-- java8版本导致javadoc打包编译失败时候，添加-->
-<!--            <properties>-->
-<!--                <javadoc.opts>-Xdoclint:none</javadoc.opts>-->
-<!--                <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>-->
-<!--            </properties>-->
         </profile>
 
         <profile>
@@ -56,10 +70,10 @@
                     <plugin>
                         <groupId>org.sonatype.plugins</groupId>
                         <artifactId>nexus-staging-maven-plugin</artifactId>
-                        <version>1.6.3</version>
+                        <version>1.6.7</version>
                         <extensions>true</extensions>
                         <configuration>
-                            <serverId>ossrh</serverId>
+                            <serverId>repository</serverId>
                             <nexusUrl>https://oss.sonatype.org/</nexusUrl>
                             <autoReleaseAfterClose>true</autoReleaseAfterClose>
                         </configuration>
@@ -82,6 +96,7 @@
                         <configuration>
                             <source>1.8</source>
                             <target>1.8</target>
+                            <encoding>utf-8</encoding>
                         </configuration>
                     </plugin>
                     <plugin>
@@ -122,10 +137,6 @@
                                 <goals>
                                     <goal>jar</goal>
                                 </goals>
-                                <!-- java8版本导致javadoc打包编译失败时候，添加-->
-<!--                                <configuration>-->
-<!--                                    <additionalparam>${javadoc.opts}</additionalparam>-->
-<!--                                </configuration>-->
                             </execution>
                         </executions>
                     </plugin>
@@ -133,19 +144,8 @@
             </build>
         </profile>
     </profiles>
-    <licenses>
-        <license>
-            <name>The Apache Software License, Version 2.0</name>
-            <url>http://www.apache.org/licenses/LICENSE-2.0.txt</url>
-            <distribution>repo</distribution>
-        </license>
-    </licenses>
 
-    <scm>
-        <url>https://github.com/MaYP1998/TaskExecutor.git</url>
-        <connection>https://github.com/MaYP1998/TaskExecutor.git</connection>
-        <developerConnection>https://github.com/MaYP1998/TaskExecutor.git</developerConnection>
-    </scm>
+
 
     <developers>
         <developer>
@@ -156,10 +156,7 @@
     </developers>
 
 
-    <!--2 -->
-
     <properties>
-        <java.version>1.8</java.version>
         <!-- 文件拷贝时的编码 -->
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
